What I've done so far for image analysis
Can import the image and find centres of each bright spot in the pattern. A major issue at the moment is that the threshold and size variables must be carefully optimized to account for noise in the image and ensure only the correct patterns are identified. Currently any incorrectly identified spot must be deleted manually. 

Working on code which will align images correctly and account for any tilt in the camera/original image.